Transaction.php: don't lock the transaction row in the transactions table
This change allows more than one request to be handled at the same time
for a given transaction. On the other hand:
* Single operation per resource must be assured separately.
* Single transaction-level operation at a given time must be assured
  separately.
It's a good question how to organize such locking in a way that handlers
don't cause deadlocks even if they use the database but a single REST
request is atomic.
It's worth noting that the old implementation just causes deadlocks when
two parallel requests are made for the same transaction.