Only RestController can output HTTP headers and data (closes #24)
Until now data has been put to PHP output in various places which could
lead to response being send to a client prematurely. Now only the
RestController class is supposed to actually send the output. It
provides setHeader()/addHeader()/setOutput() allowing other code to
impact what will be send.

For all outputs but the one generated by the MetadataReadOnly class the
change was straightforward - instead of echo-ing the response they just
generate a string and set it using RestController::setOutput() (and a
simple util/OutputFile wrapper is used to do the same for whole files).

MetadataReadOnly required a slightly more complex approach. It must
release the database connection early so its output has to be stored
somewhere and hopefully not just as a string in memory (because output
can be really huge here). This has been solved by introducing an
internal php://temp stream which automagically stores data in memory
until they are small (PHP default is 2MB) and uses a temporary file when
they grow bigger.