Adjust AMQP handlers handling
MQP handlers are now expected to return a JSON object with at least
a `status` property. If status is 0, the handlers controller tries to
parse the returned metadata from the `metadata` field. If the status is
non-zero, an exception is raised with the code equal to the status value
and the message taken from the `message` property.