Resource: tune resource creation error handling
Handle resoure creation failure in the same way both in the
postCollection() and postCollectionMetadata() methods. If the resource
removal fails on a foreign key check, log just a short error message and
in the response include the information about the original error which
cancelled the resource creation. If the resource removal fails for other
reason, log the full error message but in the response include the
information about the original error which cancelled the resource creation.