Database schema - create gin(full_text_search.raw) index
The full text search based on segments does not work for phrases being a
segment part. This particularly affect searches like parts of an URL
path, a filename extension, etc. Such searches could be easily performed
though using the "raw ILIKE '%phrase%'" just for that we need a gin
index on the raw column.